floating point arithmetic is hell & cuda is weird
floating point arithmetic is non-associative, so warp_fold results will differ subtly, usually the ulp differences will be less than 1, and will disappear, but if it doesn't and if there is a 1 ulp difference between threads, and the result for some threads is  at the finish threshold exactly, and for others is 1 ulp less than the threshold, then everything goes to hell. Then to make it even worse is that the cuda threads will behave insanely, and sometimes exit early or continue along as if it didn't happen. To make it worse, in the case of an early exit it will then deadlock, and sometimes crash with an unknown error